\name{swissMrP-package}
\alias{swissMrP-package}
\docType{package}
\title{
Multilevel Regression and Post-Stratification (MrP) for Switzerland 
}
\description{
This package provides a number of useful functions to employ MrP for small area prediction in Switzerland. Based on a hierarchical model (based itself on survey data) one can derive cantonal preference measures. The package allows to automatize the prediction and post-stratification steps. It further provides adequate print, summary, map, and plot functions for objects of its class.
}
\details{
\tabular{ll}{
Package: \tab swissMrP\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2018-26-05\cr
License: \tab GLP-2\cr
}

}
\author{
Lucas Leemann \email{lleemann@gmail.com}
}
\references{Gelman, Andrew and Thomas C. Little. 1997. Poststratification Into Many Categories Using Hierarchical Logistic Regression. \emph{Survey Research} 23:127-135.
	
	Jeffrey Lax and Justin Phillips. 2009. How Should We Estimate Public Opinion in The States? \emph{American Journal of Political Science} 53 (1), 107-121.
	
	 Leemann, Lucas and Fabio Wasserfallen. 2016. The Democratic Effect of Direct Democracy. \emph{American Political Science Review} 110(4): 750-762.
}
\note{This package was originally written for students of the University of Zurich (Fall 2013). The functions and data make this package a useful resource for anybody implementing MrP on Swiss data.
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% R documentation directory.
\keyword{MrP}
\keyword{poststratification}
\keyword{prediction}
\keyword{Estimating Public Opinion}% __ONLY ONE__ keyword per line